app_heroku/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
pandas>=1.1.3
Flask==1.1.2
Pillow==8.2.0
torch==1.8.1+cpu
torchvision==0.9.1+cpu